+++ src/TypeScript/EslintBridgeClient/EslintBridgeClient.cs
@@ -41,7 +41,7 @@ namespace SonarLint.VisualStudio.TypeScript.EslintBridgeClient
         /// <summary>
         /// Analyzes the specified file and returns the detected issues.
         /// </summary>
-        Task<AnalysisResponse> Analyze(string filePath, string tsConfigFilePath, CancellationToken cancellationToken);
+        Task<JsTsAnalysisResponse> Analyze(string filePath, string tsConfigFilePath, CancellationToken cancellationToken);
 
         /// <summary>
         /// Closes running eslint-bridge server.
@@ -98,7 +98,7 @@ namespace SonarLint.VisualStudio.TypeScript.EslintBridgeClient
             }
         }
 
-        public async Task<AnalysisResponse> Analyze(string filePath, string tsConfigFilePath, CancellationToken cancellationToken)
+        public async Task<JsTsAnalysisResponse> Analyze(string filePath, string tsConfigFilePath, CancellationToken cancellationToken)
         {
             var tsConfigFilePaths = tsConfigFilePath == null ? Array.Empty<string>() : new[] { tsConfigFilePath };
 
@@ -117,7 +117,7 @@ namespace SonarLint.VisualStudio.TypeScript.EslintBridgeClient
                 throw new InvalidOperationException(string.Format(Resources.ERR_InvalidResponse, responseString));
             }
 
-            return JsonConvert.DeserializeObject<AnalysisResponse>(responseString);
+            return JsonConvert.DeserializeObject<JsTsAnalysisResponse>(responseString);
         }
 
         public async Task Close()
